GG-37709 Use DistributedThinClientConfiguration instead
